pipelines/flows/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
apprise==1.9.3
asgi-lifespan==2.1.0
asyncpg==0.30.0
attrs==25.3.0
boto3==1.39.9
botocore==1.39.9
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coolname==2.2.0
cryptography==45.0.5
dateparser==1.2.2
docker==7.1.0
durationpy==0.10
exceptiongroup==1.3.0
fastapi==0.116.1
frozenlist==1.7.0
fsspec==2025.7.0
google-auth==2.40.3
graphviz==0.21
greenlet==3.2.3
griffe==1.7.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
iso8601==2.1.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kopf==1.38.0
kubernetes==33.1.0
kubernetes_asyncio==32.3.2
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.3
oauthlib==3.3.1
opentelemetry-api==1.35.0
orjson==3.11.0
packaging==25.0
pathspec==0.12.1
prefect==3.4.9
prefect-kubernetes==0.6.1
prometheus_client==0.22.1
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-slugify==8.0.4
python-socks==2.7.1
pytz==2025.2
pywin32==311
PyYAML==6.0.2
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.13.1
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uv==0.8.0
uvicorn==0.35.0
websocket-client==1.8.0
websockets==15.0.1
whenever==0.8.7
yarl==1.20.1
zipp==3.23.0
